Added Readme.md file

+++ app/Http/Controllers/Admin/LandlordsController.php
@@ -7,10 +7,15 @@ use Illuminate\Http\Request;
 use App\Models\Admin;
 use App\Models\Property;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Auth;
 
 class LandlordsController extends Controller
 {
     public function landlords(){
+        if(Auth::user()->role != '0'){
+            Toastr::error('You are not authorized to perform this action');
+            return redirect()->back();
+        }
         $landlords = Admin::latest()->get();
         return view('admin.pages.landlords', compact('landlords'));
     }
@@ -22,6 +27,10 @@ class LandlordsController extends Controller
     }
 
     public function adminList(){
+        if(Auth::user()->role != '0'){
+            Toastr::error('You are not authorized to perform this action');
+            return redirect()->back();
+        }
         $admins = Admin::where('role', '1')->get();
         return view('admin.pages.admin-list', compact('admins'));
  
@@ -29,11 +38,19 @@ class LandlordsController extends Controller
     }
 
     public function adminCreate(){
+        if(Auth::user()->role != '0'){
+            Toastr::error('You are not authorized to perform this action');
+            return redirect()->back();
+        }
 
         return view('admin.pages.admin-create');
     }
 
     public function adminStore(Request $request){
+        if(Auth::user()->role != '0'){
+            Toastr::error('You are not authorized to perform this action');
+            return redirect()->back();
+        }
         $validated = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:admins'],
@@ -56,6 +73,12 @@ class LandlordsController extends Controller
 
 
     public function adminDelete($id){
+
+        if(Auth::user()->role != '0'){
+            Toastr::error('You are not authorized to perform this action');
+            return redirect()->back();
+        }
+
         $admin = Admin::find($id);
         $admin->delete();
         Toastr::success('Admin Deleted Successfully');
